+++ POS/POS.Logic/Interfaces/Data/IReportEventsRepository.cs
@@ -1,0 +1,11 @@
+﻿using POS.Core.Reports;
+using System;
+using System.Threading.Tasks;
+
+namespace POS.Core.Interfaces.Data
+{
+    public interface IReportEventsRepository
+    {
+        DateTime? GetLastSuccessRunDate(string reportName, int userId);
+    }
+}