+++ POS/POS.Logic/Reports/ReportEventType.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace POS.Core.Reports
+{
+    public enum ReportEventType
+    {
+        ReportExecutedSuccess = 4001,
+        ReportExecutedFailed = 4002
+    }
+}